requirements.txt
tilly
setuptools
GitPython
sqlite-utils>=3.2
beautifulsoup4
httpx
# datasette>=0.65.1
# datasette-atom>=0.7
# datasette-publish-fly
# datasette-template-sql>=1.0.2
# datasette-graphql
# datasette-block-robots
# datasette-sitemap>=1.0
# s3-credentials
# shot-scraper
# openai-to-sqlite>=0.4.2
# sqlite-utils-sqlite-vec
# datasette-sqlite-vec